+++ scripts/automation/tests/verify_packages.py
@@ -47,7 +47,7 @@ def build_package(path_to_package, dist_dir):
     print_heading('Building {}'.format(path_to_package))
     path_to_setup = os.path.join(path_to_package, 'setup.py')
     set_version(path_to_setup)
-    cmd_success = exec_command('python setup.py sdist -d {0} bdist_wheel -d {0}'.format(dist_dir), cwd=path_to_package)
+    cmd_success = exec_command('python setup.py bdist_wheel -d {0}'.format(dist_dir), cwd=path_to_package)
     if not cmd_success:
         print_heading('Error building {}!'.format(path_to_package), f=sys.stderr)
         sys.exit(1)