+++ src/command_modules/azure-cli-lab/azure/cli/command_modules/lab/commands.py
@@ -11,12 +11,14 @@ from ._client_factory import (get_devtestlabs_virtual_machine_operation,
                               get_devtestlabs_lab_operation,
                               get_devtestlabs_virtual_network_operation,
                               get_devtestlabs_formula_operation,
-                              get_devtestlabs_secret_operation)
+                              get_devtestlabs_secret_operation,
+                              get_devtestlabs_environment_operation,
+                              get_devtestlabs_arm_template_operation)
 from azure.cli.core.sdk.util import (ServiceGroup, create_service_adapter)
 
 
 custom_path = 'azure.cli.command_modules.lab.custom'
-mgmt_operations_path = 'azure.cli.command_modules.lab.sdk.devtestlabs.operations.{}'
+mgmt_operations_path = 'azure.mgmt.devtestlabs.operations.{}'
 
 
 def _export_artifacts(formula):
@@ -39,15 +41,14 @@ custom_operations = create_service_adapter(custom_path)
 
 # Virtual Machine Operations Commands
 virtual_machine_operations = create_service_adapter(
-    mgmt_operations_path.format('virtual_machine_operations'),
-    'VirtualMachineOperations')
+    mgmt_operations_path.format('virtual_machines_operations'),
+    'VirtualMachinesOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_virtual_machine_operation,
                   virtual_machine_operations) as s:
     with s.group('lab vm') as c:
-        c.command('show', 'get_resource')
-        c.command('list', 'list')
-        c.command('delete', 'delete_resource')
+        c.command('show', 'get')
+        c.command('delete', 'delete')
         c.command('start', 'start')
         c.command('stop', 'stop')
         c.command('apply-artifacts', 'apply_artifacts')
@@ -64,32 +65,32 @@ with ServiceGroup(__name__, get_devtestlabs_lab_operation,
     with s.group('lab vm') as c:
         c.command('create', 'create_lab_vm')
 
-lab_operations = create_service_adapter(mgmt_operations_path.format('lab_operations'),
-                                        'LabOperations')
+lab_operations = create_service_adapter(mgmt_operations_path.format('labs_operations'),
+                                        'LabsOperations')
 
 # Lab Operations Commands
 with ServiceGroup(__name__, get_devtestlabs_lab_operation,
                   lab_operations) as s:
     with s.group('lab') as c:
-        c.command('get', 'get_resource')
-        c.command('delete', 'delete_resource')
+        c.command('get', 'get')
+        c.command('delete', 'delete')
 
 # Custom Image Operations Commands
 custom_image_operations = create_service_adapter(
-    mgmt_operations_path.format('custom_image_operations'),
-    'CustomImageOperations')
+    mgmt_operations_path.format('custom_images_operations'),
+    'CustomImagesOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_custom_image_operation,
                   custom_image_operations) as s:
     with s.group('lab custom-image') as c:
-        c.command('show', 'get_resource')
+        c.command('show', 'get')
         c.command('list', 'list')
-        c.command('delete', 'delete_resource')
+        c.command('delete', 'delete')
 
 # Gallery Image Operations Commands
 gallery_image_operations = create_service_adapter(
-    mgmt_operations_path.format('gallery_image_operations'),
-    'GalleryImageOperations')
+    mgmt_operations_path.format('gallery_images_operations'),
+    'GalleryImagesOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_gallery_image_operation,
                   gallery_image_operations) as s:
@@ -98,8 +99,8 @@ with ServiceGroup(__name__, get_devtestlabs_gallery_image_operation,
 
 # Artifact Operations Commands
 artifact_operations = create_service_adapter(
-    mgmt_operations_path.format('artifact_operations'),
-    'ArtifactOperations')
+    mgmt_operations_path.format('artifacts_operations'),
+    'ArtifactsOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_artifact_operation,
                   artifact_operations) as s:
@@ -108,48 +109,83 @@ with ServiceGroup(__name__, get_devtestlabs_artifact_operation,
 
 # Artifact Source Operations Commands
 artifact_source_operations = create_service_adapter(
-    mgmt_operations_path.format('artifact_source_operations'),
-    'ArtifactSourceOperations')
+    mgmt_operations_path.format('artifact_sources_operations'),
+    'ArtifactSourcesOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_artifact_source_operation,
                   artifact_source_operations) as s:
     with s.group('lab artifact-source') as c:
         c.command('list', 'list')
-        c.command('get', 'get_resource')
+        c.command('show', 'get')
 
 # Virtual Network Operations Commands
 virtual_network_operations = create_service_adapter(
-    mgmt_operations_path.format('virtual_network_operations'),
-    'VirtualNetworkOperations')
+    mgmt_operations_path.format('virtual_networks_operations'),
+    'VirtualNetworksOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_virtual_network_operation,
                   virtual_network_operations) as s:
     with s.group('lab vnet') as c:
         c.command('list', 'list')
-        c.command('get', 'get_resource')
+        c.command('get', 'get')
 
 # Formula Operations Commands
 formula_operations = create_service_adapter(
-    mgmt_operations_path.format('formula_operations'),
-    'FormulaOperations')
+    mgmt_operations_path.format('formulas_operations'),
+    'FormulasOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_formula_operation,
                   formula_operations) as s:
     with s.group('lab formula') as c:
-        c.command('show', 'get_resource')
+        c.command('show', 'get')
         c.command('list', 'list')
-        c.command('delete', 'delete_resource')
-        c.command('export-artifacts', 'get_resource', transform=_export_artifacts)
+        c.command('delete', 'delete')
+        c.command('export-artifacts', 'get', transform=_export_artifacts)
 
 # Secret Operations Commands
 secret_operations = create_service_adapter(
-    mgmt_operations_path.format('secret_operations'),
-    'SecretOperations')
+    mgmt_operations_path.format('secrets_operations'),
+    'SecretsOperations')
 
 with ServiceGroup(__name__, get_devtestlabs_secret_operation,
                   secret_operations) as s:
     with s.group('lab secret') as c:
-        c.command('set', 'create_or_update_resource')
-        c.command('show', 'get_resource')
+        c.command('set', 'create_or_update')
+        c.command('show', 'get')
         c.command('list', 'list')
-        c.command('delete', 'delete_resource')
+        c.command('delete', 'delete')
+
+# Environment Operations Commands
+environment_operations = create_service_adapter(
+    mgmt_operations_path.format('environments_operations'),
+    'EnvironmentsOperations')
+
+with ServiceGroup(__name__, get_devtestlabs_environment_operation,
+                  environment_operations) as s:
+    with s.group('lab environment') as c:
+        c.command('show', 'get')
+        c.command('list', 'list')
+        c.command('delete', 'delete')
+        c.command('create', 'create_or_update')
+
+# Environment Operations Custom Commands
+with ServiceGroup(__name__, get_devtestlabs_environment_operation,
+                  custom_operations) as s:
+    with s.group('lab environment') as c:
+        c.command('create', 'create_environment')
+
+# ARM Templates Operations Commands
+arm_template_operations = create_service_adapter(
+    mgmt_operations_path.format('arm_templates_operations'),
+    'ArmTemplatesOperations')
+
+with ServiceGroup(__name__, get_devtestlabs_arm_template_operation,
+                  arm_template_operations) as s:
+    with s.group('lab arm-template') as c:
+        c.command('list', 'list')
+
+# ARM Templates Operations Custom Commands
+with ServiceGroup(__name__, get_devtestlabs_arm_template_operation,
+                  custom_operations) as s:
+    with s.group('lab arm-template') as c:
+        c.command('show', 'show_arm_template')